error mapping struct and boilerplate map_err code
