+++ crates/beerus-core/Cargo.toml
@@ -24,7 +24,6 @@ toml = "0.7.3"
 
 [dev-dependencies]
 httpmock = "0.6.7"
-temp-env = "0.3.1"
 serial_test = "1.0.0"
 
 [dev-dependencies.cargo-husky]
