+++ crates/beerus-core/tests/common/data/goerli.toml
@@ -3,3 +3,4 @@
     ethereum_execution_rpc = "https://eth-goerli.g.alchemy.com/v2/XXXXX"
     starknet_rpc = "https://starknet-goerli.infura.io/v3/XXXXX"
     data_dir = "~/.beerus/tmp"
+
