+++ crates/beerus-rpc/Cargo.toml
@@ -20,6 +20,8 @@ tokio = { version = "1.8.1", features = ["full"] }
 serde_json = "1.0.64"
 env_logger = "0.10.0"
 anyhow = "1.0.69"
-wiremock = "0.5.17"
 serde = "1.0.152"
 reqwest = "0.11.13"
+
+[dev-dependencies]
+wiremock = "0.5.17"
