+++ crates/beerus-rpc/tests/common/data/test.toml
@@ -1,0 +1,5 @@
+    ethereum_network = "goerli"
+    ethereum_consensus_rpc = ""
+    ethereum_execution_rpc = ""
+    starknet_rpc = ""
+    data_dir = ""
