+++ crates/beerus-rpc/tests/common/mod.rs
@@ -5,10 +5,11 @@ use beerus_core::{
         starknet::StarkNetLightClientImpl,
     },
 };
-use beerus_rpc::server::BeerusRpc;
+use beerus_rpc::BeerusRpc;
 use reqwest::StatusCode;
 use serde::{Deserialize, Serialize};
 use starknet::providers::jsonrpc::models::{BlockId, BlockTag};
+use std::path::PathBuf;
 use wiremock::{
     matchers::{body_json, method},
     Mock, MockServer, ResponseTemplate,
@@ -30,7 +31,7 @@ pub struct EthJsonRpcResponse<StarknetParams> {
 }
 
 impl<'a, StarknetParams> StarknetRpcBaseData<'a, StarknetParams> {
-    pub const fn stark_block_number(params: StarknetParams) -> Self {
+    pub const fn block_number(params: StarknetParams) -> Self {
         Self {
             id: 1,
             jsonrpc: "2.0",
@@ -57,9 +58,8 @@ pub async fn setup_wiremock() -> String {
 }
 
 pub async fn setup_beerus_rpc() -> BeerusRpc {
-    let mock_starknet_rpc = setup_wiremock().await;
-    set_mandatory_envs(mock_starknet_rpc);
-    let config = Config::from_env();
+    let mut config = Config::from_file(&PathBuf::from("tests/common/data/test.toml"));
+    config.starknet_rpc = setup_wiremock().await;
 
     let ethereum_lightclient = MockEthereumLightClient::new();
     let starknet_lightclient = StarkNetLightClientImpl::new(&config).unwrap();
@@ -72,17 +72,9 @@ pub async fn setup_beerus_rpc() -> BeerusRpc {
     BeerusRpc::new(beerus_client)
 }
 
-fn set_mandatory_envs(starknet_rpc: String) {
-    Config::clean_env();
-    std::env::set_var("ETHEREUM_CONSENSUS_RPC_URL", "");
-    std::env::set_var("ETHEREUM_EXECUTION_RPC_URL", "");
-    std::env::set_var("STARKNET_RPC_URL", starknet_rpc);
-    std::env::set_var("DATA_DIR", "");
-}
-
 fn mock_block_number() -> Mock {
     Mock::given(method("POST"))
-        .and(body_json(StarknetRpcBaseData::stark_block_number(())))
+        .and(body_json(StarknetRpcBaseData::block_number(())))
         .respond_with(response_template_with_status(StatusCode::OK).set_body_raw(
             include_str!("data/starknet_blockNumber.json"),
             "application/json",
